+++ src/main/kotlin/net/hennabatch/vrcneochi/listener/ResetListener.kt
@@ -4,7 +4,7 @@ import net.hennabatch.vrcneochi.logger.logger
 import net.hennabatch.vrcneochi.parameter.VRCParameter
 import java.util.concurrent.TransferQueue
 
-class ResetListener(private val queue: TransferQueue<Float>) : VRCOSCParameterListener<Boolean>() {
+class ResetListener(private val queue: TransferQueue<Int>) : VRCOSCParameterListener<Boolean>() {
 
     override fun exAddress(): String {
         return "/avatar/parameters/reset_sleep"
@@ -13,7 +13,7 @@ class ResetListener(private val queue: TransferQueue<Float>) : VRCOSCParameterLi
     override fun recieveParameter(exParameter: VRCParameter<Boolean>) {
         if(exParameter.parameter) {
             //リセットを送出
-            queue.add(-1.0f)
+            queue.add(-1)
             logger.info("reset timer")
         }
     }
